+++ src/components/common/CodeBlock.vue
@@ -11,12 +11,10 @@
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted, ref, nextTick } from 'vue'
+import { computed } from 'vue'
 import composables from '@/composables'
 import { requestSampleConfigs } from '@/constants'
 import type { LanguageCode } from '@/types/request-languages'
-
-
 
 const props = defineProps({
   code: {
@@ -41,8 +39,6 @@ const highlightedCode = computed(():string => {
   }
   return ''
 })
-
-
 </script>
 
 <style lang="scss" scoped>
@@ -54,6 +50,8 @@ const highlightedCode = computed(():string => {
     font-weight: var(--kui-font-weight-regular, $kui-font-weight-regular);
     line-height: var(--kui-line-height-30, $kui-line-height-30);
     margin: var(--kui-space-0, $kui-space-0);
+    max-height: 300px;
+    overflow-y: auto;
     padding: var(--kui-space-40, $kui-space-40);
     white-space: break-spaces;
 
@@ -63,9 +61,8 @@ const highlightedCode = computed(():string => {
       word-wrap: break-word;
     }
 
-    $codeblock-line-count-width: 20px;
+    $codeblock-line-count-width: 4ch;
     $codeblock-line-gap: 1px;
-
 
     span.line {
       counter-increment: codeblock-line;
@@ -78,7 +75,7 @@ const highlightedCode = computed(():string => {
       position: relative;
       word-break: break-all;
 
-      &:after {
+      &::after {
         background-color: var(--kui-color-background-neutral-weakest, $kui-color-background-neutral-weakest);
         bottom: 0;
         color: var(--kui-color-text-neutral, $kui-color-text-neutral-weak);